requirements.txt
numpy>=1.24
pandas>=2.0
opencv-python>=4.8
Pillow>=10.0
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.3
tensorflow>=2.13
